Move the algorithm into a new file
